+++ Backend/talkative/src/main/java/com/talkative/repository/ChatRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface ChatRepository extends JpaRepository<Chat, Integer> {
+public interface ChatRepository extends JpaRepository<Chat, Long> {
 
 	@Query("select c from Chat c join c.users u where u.id =:userId")
 	public List<Chat> findChatByUserId(@Param("userId") Long userId);
